requirements.txt
discord.py
python-dotenv
aiohttp
python-docx
google-auth
google-api-python-client
